refactoring of internal folder
