+++ api_files/request_models/card_groups_model.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+# * https://fastapi.tiangolo.com/tutorial/body/#__tabbed_2_1
+class CardGroups(BaseModel):
+    set: str
+    id: str
+    group: str
